% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{confidence}
\alias{confidence}
\title{confidence interval of mean using t-test}
\usage{
confidence(x, f = twoDp, ...)
}
\arguments{
\item{x}{a numeric or vector of numerics}

\item{f}{a formatting function}
}
\description{
confidence interval of mean using t-test
}
\examples{
df \%>\% mutate(display = confidence(x))
}
\keyword{format}
