% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{countWithOtherCategory}
\alias{countWithOtherCategory}
\title{Truncate a count}
\usage{
countWithOtherCategory(df, n, countVar, label = "Other...")
}
\arguments{
\item{df}{- a grouped df}

\item{n}{the number of rows to keep}

\item{countVar}{- the column containing the count, or in which the count is to be written}

\item{label}{-  the name of the "Other..." category}
}
\description{
Truncate the results of a count using an "other..." category
}
