% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{dayXaxis}
\alias{dayXaxis}
\title{A standard X axis with hour from midnight}
\usage{
dayXaxis()
}
\description{
A standard X axis with hour from midnight
}
\examples{
ggplot()+dayXaxis()
}
\keyword{axes}
