% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{defaultFigureLayout}
\alias{defaultFigureLayout}
\title{A ggplot theme with standard print publication defaults with a focus on 10pt Arial as the default font}
\usage{
defaultFigureLayout(...)
}
\description{
A ggplot theme with standard print publication defaults with a focus on 10pt Arial as the default font
}
\examples{
library(ggplot2)
library(standardPrintOutput)
theme_set(defaultFigureLayout())
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
}
\keyword{plot}
\keyword{theme}
