% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils.R
\name{defaultGraphLayout}
\alias{defaultGraphLayout}
\title{A default dot theme with standard print publication defaults with a focus on 10pt Arial as the default font}
\usage{
defaultGraphLayout(graph, ...)
}
\description{
A default dot theme with standard print publication defaults with a focus on 10pt Arial as the default font
}
\examples{
library(ggplot2)
library(standardPrintOutput)
theme_set(defaultFigureLayout())
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
}
\keyword{DiagrammeR}
\keyword{theme}
