% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{labelInPoints}
\alias{labelInPoints}
\title{A fix geom_text sizes to be in the correct size for pdf export}
\usage{
labelInPoints(pts)
}
\arguments{
\item{pts}{size of text in points}
}
\description{
A fix geom_text sizes to be in the correct size for pdf export
}
\examples{
...+geom_label_repel(size=labelInPoints(10))
}
\keyword{graph}
\keyword{layout}
