% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtableUtils.R
\name{mergeCells}
\alias{mergeCells}
\title{prepare a huxtable with cells merged according to grouped colums}
\usage{
mergeCells(labelledDataFrame)
}
\arguments{
\item{labelledDataFrame}{e.g. select(dataframe,"col 1 title"=col1)}
}
\description{
prepare a huxtable with cells merged according to grouped colums
}
\examples{
setwd(tempdir())
mergeCells(iris \%>\% group_by(Species,Petal.Width) \%>\% summarise(count = n()))
mergeCells(mtcars \%>\% rownames_to_column() \%>\% group_by(gear,carb))
mtcars \%>\% rownames_to_column() \%>\% group_by(gear,carb) \%>\% mergeCells() \%>\% saveTable("cars")
}
\keyword{huxtable}
