% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{monthsXaxis}
\alias{monthsXaxis}
\title{A standard X axis of months of year where data is a 'MM-DD' formatted date object}
\usage{
monthsXaxis()
}
\description{
A standard X axis of months of year where data is a 'MM-DD' formatted date object
}
\examples{
ggplot()+dayXaxis()
}
\keyword{axes}
