% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{monthsXaxis2}
\alias{monthsXaxis2}
\title{A standard X axis of months of year where data is a 'MM' formatted date object}
\usage{
monthsXaxis2(plot)
}
\description{
A standard X axis of months of year where data is a 'MM' formatted date object
}
\examples{
ggplot()+monthXaxis2()
}
\keyword{axes}
