% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{narrowAndTall}
\alias{narrowAndTall}
\title{A ggplot theme modification for graphs that need to be narrower and taller than defaults}
\usage{
narrowAndTall()
}
\description{
A ggplot theme modification for graphs that need to be narrower and taller than defaults
}
\examples{
ggplot()+narrowAndTall()
}
\keyword{graph}
\keyword{layout}
