% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveFigure}
\alias{saveFigure}
\title{A standard plot for publication}
\usage{
saveFigure(filename, maxWidth, maxHeight,
  aspectRatio = maxWidth/maxHeight, plot = last_plot())
}
\arguments{
\item{filename}{base of target filename (excuding extension).}

\item{maxWidth}{maximum width in inches}

\item{maxHeight}{maximum height in inches}

\item{aspectRatio}{defaults to maxWidth/maxHeight}

\item{plot}{a GGplot object or none}
}
\description{
A standard plot for publication
}
\examples{
saveFigure("the_filename",maxWidth=4,maxHeight=4,plot=ggplot())
}
\keyword{plot}
