% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveFigureLandscape}
\alias{saveFigureLandscape}
\title{Saves standard plots for publication either as EPS, PNG with 300dpi or PDF
Allows specific maximum dimensions with an optional target aspect ratio.
If no aspect ratio supplied then plot will be shaped to occupy whole area}
\usage{
saveFigureLandscape(plot = last_plot(), filename, maxWidth, maxHeight,
  aspectRatio = maxHeight/maxWidth)
}
\arguments{
\item{plot}{a GGplot object or none}

\item{filename}{base of target filename (excuding extension).}

\item{maxWidth}{maximum width in inches (of page)}

\item{maxHeight}{maximum height in inches (of page)}

\item{aspectRatio}{defaults to maxHeight/maxWidth}
}
\description{
Saves standard plots for publication either as EPS, PNG with 300dpi or PDF
Allows specific maximum dimensions with an optional target aspect ratio.
If no aspect ratio supplied then plot will be shaped to occupy whole area
}
\examples{
setwd(tempdir())
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
saveFigure(filename="the_filename",maxWidth=4,maxHeight=4)
}
\keyword{plot}
