% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveFullPageFigure}
\alias{saveFullPageFigure}
\title{A standard 6x8 inch plot size for a full page}
\usage{
saveFullPageFigure(filename, ...)
}
\arguments{
\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveThesis()}

\item{plot}{a GGplot object or none}
}
\description{
A standard 6x8 inch plot size for a full page
}
\examples{
saveFullPageFigure("the_filename",ggplot())
}
\keyword{axes}
