% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils.R
\name{saveFullPageGraph}
\alias{saveFullPageGraph}
\title{A standard 6x8 inch plot size for a full page
This may be combined with narrowAndTall() to maximise plotting area}
\usage{
saveFullPageGraph(graph, filename, ...)
}
\arguments{
\item{graph}{a DiagrammeR::graph object}

\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveFigure()}
}
\description{
A standard 6x8 inch plot size for a full page
This may be combined with narrowAndTall() to maximise plotting area
}
\examples{
setwd(tempdir())
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
saveFullPageFigure(filename="the_filename")
}
\keyword{axes}
