% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils.R
\name{saveFullPageGraphLandscape}
\alias{saveFullPageGraphLandscape}
\title{A standard 6x8 inch plot size for a full page in landscape format}
\usage{
saveFullPageGraphLandscape(graph, filename, ...)
}
\arguments{
\item{graph}{a DiagrammeR::graph object}

\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveFigure()}
}
\description{
A standard 6x8 inch plot size for a full page in landscape format
}
\examples{
setwd(tempdir())
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
saveFullPageLandscapeFigure(filename="the_filename")
}
\keyword{axes}
