% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils.R
\name{saveGraph}
\alias{saveGraph}
\title{Saves standard graphs for publication either as EPS, PNG with 300dpi or PDF
Allows specific maximum dimensions with an optional target aspect ratio.
If no aspect ratio supplied then plot will be shaped to occupy whole area}
\usage{
saveGraph(graph, filename, maxWidth, maxHeight, engine = "dot")
}
\arguments{
\item{graph}{a DiagrammeR::graph object}

\item{filename}{base of target filename (excuding extension).}

\item{maxWidth}{maximum width in inches}

\item{maxHeight}{maximum height in inches}
}
\description{
Saves standard graphs for publication either as EPS, PNG with 300dpi or PDF
Allows specific maximum dimensions with an optional target aspect ratio.
If no aspect ratio supplied then plot will be shaped to occupy whole area
}
\examples{
setwd(tempdir())
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
saveFigure(filename="the_filename",maxWidth=4,maxHeight=4)
}
\keyword{plot}
