% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveHalfPageFigure}
\alias{saveHalfPageFigure}
\title{A standard max 6x4 inch plot size for a half page}
\usage{
saveHalfPageFigure(filename, ...)
}
\arguments{
\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveThesis()}

\item{plot}{a GGplot object or none}
}
\description{
A standard max 6x4 inch plot size for a half page
}
\examples{
saveHalfPageFigure("the_filename",ggplot())
}
\keyword{axes}
