% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveQuarterPageFigure}
\alias{saveQuarterPageFigure}
\title{A standard max 6x2 plot size for a quarter page}
\usage{
saveQuarterPageFigure(plot, filename, ...)
}
\arguments{
\item{plot}{a GGplot object or none}

\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveFigure()}
}
\description{
A standard max 6x2 plot size for a quarter page
}
\examples{
setwd(tempdir())
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
saveQuarterPageFigure(filename="the_filename")
}
\keyword{axes}
