% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils.R
\name{saveSixthPageGraph}
\alias{saveSixthPageGraph}
\title{A standard max 3x3 plot size for a smaller plot
maybe combined with narrowAndTall()}
\usage{
saveSixthPageGraph(graph, filename, ...)
}
\arguments{
\item{graph}{a DiagrammeR::graph object}

\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveFigure()}
}
\description{
A standard max 3x3 plot size for a smaller plot
maybe combined with narrowAndTall()
}
\examples{
setwd(tempdir())
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
saveSixthPageFigure(filename="the_filename")
}
\keyword{axes}
