% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtableUtils.R
\name{saveTableLandscape}
\alias{saveTableLandscape}
\title{save labelled dataframe to html and pdf file silently in rotated format for a landscape page}
\usage{
saveTableLandscape(
  labelledDataframe,
  filename,
  pageLength = 8,
  defaultFontSize = 10,
  tableWidth = NULL,
  colWidths = NULL
)
}
\arguments{
\item{filename}{file of desired output without extension.}

\item{pageLength}{maximum length of the page in inches (8)}

\item{defaultFontSize}{default size of font in points (10)}

\item{tableWidth}{width of the table in inches or NULL to fit contents automatically}

\item{colWidths}{a vector of relative column widths}

\item{labelledDataFrame}{e.g. dataframe %>% select("col 1 title"=col1)}
}
\description{
save labelled dataframe to html and pdf file silently in rotated format for a landscape page
}
\examples{
setwd(tempdir())
hux(iris) \%>\% saveTableLandscape("iris")
}
\keyword{huxtable}
\keyword{webshot}
