% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveThirdPageFigure}
\alias{saveThirdPageFigure}
\title{A standard max 6x3 plot size for a third page}
\usage{
saveThirdPageFigure(filename, ...)
}
\arguments{
\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveThesis()}

\item{plot}{a GGplot object or none}
}
\description{
A standard max 6x3 plot size for a third page
}
\examples{
saveThirdPageFigure("the_filename",ggplot())
}
\keyword{axes}
