% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{scale_day_of_year_discrete}
\alias{scale_day_of_year_discrete}
\title{A standard X axis of months of year where data is a 'MM-DD' formatted date object}
\usage{
scale_day_of_year_discrete()
}
\description{
A standard X axis of months of year where data is a 'MM-DD' formatted date object
}
\examples{
ggplot()+scale_day_of_year_discrete()
}
\keyword{axes}
