% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{scale_hours_continuous}
\alias{scale_hours_continuous}
\title{A standard X axis with hour from midnight}
\usage{
scale_hours_continuous()
}
\description{
A standard X axis with hour from midnight
}
\examples{
ggplot()+scale_hours_continuous()
}
\keyword{axes}
