% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{scale_month_discrete}
\alias{scale_month_discrete}
\title{A standard X axis of months of year where data is a 'MM' formatted date object}
\usage{
scale_month_discrete()
}
\description{
A standard X axis of months of year where data is a 'MM' formatted date object
}
\examples{
ggplot()+scale_month_discrete()
}
\keyword{axes}
