% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{scale_week_continuous}
\alias{scale_week_continuous}
\title{A standard X axis with hour from a sunday}
\usage{
scale_week_continuous()
}
\description{
A standard X axis with hour from a sunday
}
\examples{
ggplot()+scale_week_continuous()
}
\keyword{axes}
