% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{themeStandardOutput}
\alias{themeStandardOutput}
\title{A ggplot theme with standard print publication defaults}
\usage{
themeStandardOutput(...)
}
\description{
A ggplot theme with standard print publication defaults
}
\examples{
theme_set(themeStandardOutput())
}
\keyword{plot}
\keyword{theme}
