% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{threeDp}
\alias{threeDp}
\title{3 decimal place}
\usage{
threeDp(x, unit = "")
}
\arguments{
\item{x}{a numeric or vector of numerics}

\item{unit}{a unit specification}
}
\description{
3 decimal place
}
\examples{
df \%>\% mutate(display = threeDp(x))
df \%>\% mutate(display = threeDp(x,"mm"))
}
\keyword{format}
