% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{watermark}
\alias{watermark}
\title{A ggplot annotation layer that puts a watermark on a graph}
\usage{
watermark(show = TRUE, lab = "DRAFT")
}
\arguments{
\item{show}{a flag to allow a global level of control}

\item{lab}{a label to annotate (e.g. DRAFT)}
}
\description{
A ggplot annotation layer that puts a watermark on a graph
}
\examples{
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point() + watermark()
}
\keyword{graph}
\keyword{layout}
