% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{weekXaxis}
\alias{weekXaxis}
\title{A standard X axis with hour from a sunday}
\usage{
weekXaxis()
}
\description{
A standard X axis with hour from a sunday
}
\examples{
ggplot()+weekXaxis()
}
\keyword{axes}
