% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{narrowAndTall}
\alias{narrowAndTall}
\title{A ggplot theme modification for graphs that need to be narrower and taller than defaults}
\usage{
narrowAndTall()
}
\description{
A ggplot theme modification for graphs that need to be narrower and taller than defaults
}
\examples{
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point() + narrowAndTall()
}
\keyword{graph}
\keyword{layout}
