% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveFullPageFigure}
\alias{saveFullPageFigure}
\title{A standard 6x8 inch plot size for a full page
This may be combined with narrowAndTall() to maximise plotting area}
\usage{
saveFullPageFigure(plot = last_plot(), filename, ...)
}
\arguments{
\item{plot}{a GGplot object or none}

\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveFigure()}
}
\description{
A standard 6x8 inch plot size for a full page
This may be combined with narrowAndTall() to maximise plotting area
}
\examples{
setwd(tempdir())
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
saveFullPageFigure("the_filename")
}
\keyword{axes}
