% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveHalfPageFigure}
\alias{saveHalfPageFigure}
\title{A standard max 6x4 inch plot size for a half page}
\usage{
saveHalfPageFigure(plot = last_plot(), filename, ...)
}
\arguments{
\item{plot}{a GGplot object or none}

\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveFigure()}
}
\description{
A standard max 6x4 inch plot size for a half page
}
\examples{
setwd(tempdir())
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
saveHalfPageFigure("the_filename")
}
\keyword{axes}
