% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveThirdPageFigure}
\alias{saveThirdPageFigure}
\title{A standard max 6x3 plot size for a third page}
\usage{
saveThirdPageFigure(plot = last_plot(), filename, ...)
}
\arguments{
\item{plot}{a GGplot object or none}

\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveFigure()}
}
\description{
A standard max 6x3 plot size for a third page
}
\examples{
setwd(tempdir())
library(ggplot2)
library(standardPrintOutput)
ggplot(mtcars, aes(mpg, wt, colour=as.factor(cyl))) + geom_point()
saveThirdPageFigure(filename="the_filename")
}
\keyword{axes}
